add auth for every route
