+++ backend/src/common/types.ts
@@ -2,14 +2,18 @@ import { NextFunction, Request, Response } from "express";
 import { ValidationChain } from "express-validator";
 import { UserDB } from "./generated.js";
 
-export type AuthenticatedUserRequest = Request & {
-    user?: User
+export type AuthenticatedRequest = Request & {
+    user?: RequestUser
 }
 
-// exclude password
-export type User = Pick<UserDB, Exclude<keyof UserDB, "password">>;
+/**
+ * UserDB with password excluded
+ */
+export type RequestUser = Pick<UserDB, Exclude<keyof UserDB, "password">>;
 
-// Interface for the decoded data from the JWT
+/**
+ * Interface for the decoded data from a JWT
+ */
 export interface DecodedJwtPayload {
     user_id: string;
 }
