+++ backend/src/controllers/classController.ts
@@ -1,16 +1,17 @@
-import { Request, Response } from 'express';
+import { Response } from 'express';
 import { ClassDB } from '../common/generated.js';
+import { AuthenticatedRequest } from '../common/types.js';
 import ClassesModel from '../models/classModel.js';
 
 const classesModel = new ClassesModel();
 
-async function getAllClasses(req: Request, res: Response) {
+async function getAllClasses(req: AuthenticatedRequest, res: Response) {
 	const classes = await classesModel.getClasses();
 
 	res.status(200).json(classes);
 }
 
-async function addClass(req: Request, res: Response) {
+async function addClass(req: AuthenticatedRequest, res: Response) {
 	const newClass: ClassDB = req.body;
 	const schedules = req.body.schedules;
 
@@ -19,7 +20,7 @@ async function addClass(req: Request, res: Response) {
 	res.status(200).json(result);
 }
 
-async function updateClass(req: Request, res: Response) {
+async function updateClass(req: AuthenticatedRequest, res: Response) {
 	const class_id = Number(req.params.class_id);
 	const partialClass: Partial<ClassDB> = req.body;
 
@@ -28,7 +29,7 @@ async function updateClass(req: Request, res: Response) {
 	return res.status(200).json(result);
 }
 
-async function deleteClass(req: Request, res: Response) {
+async function deleteClass(req: AuthenticatedRequest, res: Response) {
 	const class_id = Number(req.params.class_id);
 
 	const result = await classesModel.deleteClass(class_id);
@@ -36,7 +37,7 @@ async function deleteClass(req: Request, res: Response) {
 	return res.status(200).json(result);
 }
 
-async function getClassesByDay(req: Request, res: Response) {
+async function getClassesByDay(req: AuthenticatedRequest, res: Response) {
 	const { day } = req.params;
 
 	const classes = await classesModel.getClassesByDay(day);
@@ -45,7 +46,7 @@ async function getClassesByDay(req: Request, res: Response) {
 }
 
 // get class info for a specific shift ID
-async function getClassById(req: Request, res: Response) {
+async function getClassById(req: AuthenticatedRequest, res: Response) {
 	const class_id = Number(req.params.class_id);
 
 	const class_info = await classesModel.getClassById(class_id);
@@ -53,7 +54,7 @@ async function getClassById(req: Request, res: Response) {
 	res.status(200).json(class_info);
 }
 
-async function uploadClassImage(req: Request, res: Response) {
+async function uploadClassImage(req: AuthenticatedRequest, res: Response) {
 	const class_id = Number(req.params.class_id);
 	const image = req.file!.buffer;
