+++ backend/src/controllers/imageController.ts
@@ -1,9 +1,10 @@
-import { Request, Response } from "express";
+import { Response } from "express";
+import { AuthenticatedRequest } from "../common/types.js";
 import ImageModel from "../models/imageModel.js";
 
 const imageService = new ImageModel();
 
-export async function getImage(req: Request, res: Response) {
+export async function getImage(req: AuthenticatedRequest, res: Response) {
     const { image_id } = req.params;
 
     const image = await imageService.getImage(image_id);
