+++ backend/src/controllers/instructorController.ts
@@ -1,16 +1,17 @@
-import { Request, Response } from 'express';
+import { Response } from 'express';
 import { InstructorDB } from '../common/generated.js';
+import { AuthenticatedRequest } from '../common/types.js';
 import InstructorModel from '../models/instructorModel.js';
 
 const instructorModel = new InstructorModel();
 
-async function getInstructors(req: Request, res: Response) {
+async function getInstructors(req: AuthenticatedRequest, res: Response) {
     const instructors = await instructorModel.getInstructors();
 
     res.status(200).json(instructors);
 }
 
-async function getInstructorById(req: Request, res: Response) {
+async function getInstructorById(req: AuthenticatedRequest, res: Response) {
     const { instructor_id } = req.params;
 
     const instructor = await instructorModel.getInstructorById(instructor_id);
@@ -18,7 +19,7 @@ async function getInstructorById(req: Request, res: Response) {
     res.status(200).json(instructor);
 }
 
-async function insertInstructor(req: Request, res: Response) {
+async function insertInstructor(req: AuthenticatedRequest, res: Response) {
     const instructor: InstructorDB = req.body;
 
     const result = await instructorModel.insertInstructor(instructor);
