+++ backend/src/controllers/scheduleController.ts
@@ -1,16 +1,17 @@
-import { Request, Response } from 'express';
+import { Response } from 'express';
 import { ScheduleDB } from '../common/generated.js';
+import { AuthenticatedRequest } from '../common/types.js';
 import ScheduleModel from '../models/scheduleModel.js';
 
 const scheduleModel = new ScheduleModel();
 
-async function getAllSchedules(req: Request, res: Response) {
+async function getAllSchedules(req: AuthenticatedRequest, res: Response) {
     const schedules = await scheduleModel.getAllSchedules();
     
     res.status(200).json(schedules);
 }
 
-async function getActiveSchedulesForClass(req: Request, res: Response) {
+async function getActiveSchedulesForClass(req: AuthenticatedRequest, res: Response) {
     const class_id = Number(req.params.class_id);
 
     const schedules = await scheduleModel.getActiveSchedulesForClass(class_id);
@@ -18,7 +19,7 @@ async function getActiveSchedulesForClass(req: Request, res: Response) {
     res.status(200).json(schedules);
 }
 
-async function addSchedulesToClass(req: Request, res: Response) {
+async function addSchedulesToClass(req: AuthenticatedRequest, res: Response) {
     const class_id = Number(req.params.class_id);
     const schedules: ScheduleDB[] = req.body;
 
@@ -27,7 +28,7 @@ async function addSchedulesToClass(req: Request, res: Response) {
     res.status(200).json(result);
 }
 
-async function deleteOrSoftDeleteSchedules(req: Request, res: Response) {
+async function deleteOrSoftDeleteSchedules(req: AuthenticatedRequest, res: Response) {
     const class_id = Number(req.params.class_id);
     const { schedule_ids } = req.body;
 
@@ -36,7 +37,7 @@ async function deleteOrSoftDeleteSchedules(req: Request, res: Response) {
     res.status(200).json(result);
 }
 
-async function deleteSchedulesFromClass(req: Request, res: Response) {
+async function deleteSchedulesFromClass(req: AuthenticatedRequest, res: Response) {
     const class_id = Number(req.params.class_id);
     const { schedule_ids } = req.body;
 
@@ -45,7 +46,7 @@ async function deleteSchedulesFromClass(req: Request, res: Response) {
     res.status(200).json(result);
 }
 
-async function assignVolunteersToSchedule(req: Request, res: Response) {
+async function assignVolunteersToSchedule(req: AuthenticatedRequest, res: Response) {
     const class_id = Number(req.params.class_id);
     const schedule_id = Number(req.params.schedule_id);
     const { volunteer_ids } = req.body;
@@ -55,7 +56,7 @@ async function assignVolunteersToSchedule(req: Request, res: Response) {
     res.status(200).json(result);
 }
 
-async function updateSchedulesForClass(req: Request, res: Response) {
+async function updateSchedulesForClass(req: AuthenticatedRequest, res: Response) {
     const class_id = Number(req.params.class_id);
     const schedules: ScheduleDB[] = req.body;
 
@@ -65,12 +66,8 @@ async function updateSchedulesForClass(req: Request, res: Response) {
 }
 
 export {
-    getAllSchedules,
-    getActiveSchedulesForClass,
-    deleteOrSoftDeleteSchedules,
-    deleteSchedulesFromClass,
     addSchedulesToClass,
-    assignVolunteersToSchedule,
-    updateSchedulesForClass
+    assignVolunteersToSchedule, deleteOrSoftDeleteSchedules,
+    deleteSchedulesFromClass, getActiveSchedulesForClass, getAllSchedules, updateSchedulesForClass
 };
