+++ backend/src/controllers/volunteerController.ts
@@ -1,9 +1,10 @@
-import { Request, Response } from "express";
+import { Response } from "express";
+import { AuthenticatedRequest } from "../common/types.js";
 import VolunteerModel from "../models/volunteerModel.js";
 
 const volunteerModel = new VolunteerModel();
 
-async function getVolunteerById(req: Request, res: Response) {
+async function getVolunteerById(req: AuthenticatedRequest, res: Response) {
     const { volunteer_id } = req.params;
 
     const volunteer = await volunteerModel.getVolunteerById(volunteer_id);
@@ -11,14 +12,14 @@ async function getVolunteerById(req: Request, res: Response) {
     res.status(200).json(volunteer);
 }
 
-async function getVolunteers(req: Request, res: Response) {
+async function getVolunteers(req: AuthenticatedRequest, res: Response) {
     const volunteers = await volunteerModel.getVolunteers();
 
     res.status(200).json(volunteers);
 }
 
 // Update a volunteer's profile based on the volunteer_id
-async function updateVolunteer(req: Request, res: Response) {
+async function updateVolunteer(req: AuthenticatedRequest, res: Response) {
     const { volunteer_id } = req.params;
     const volunteerData = req.body;
 
@@ -31,7 +32,7 @@ async function updateVolunteer(req: Request, res: Response) {
 }
 
 // Update a volunteer's profile based on the volunteer_id
-async function shiftCheckIn(req: Request, res: Response) {
+async function shiftCheckIn(req: AuthenticatedRequest, res: Response) {
     const { volunteerID, scheduleID, shiftDate } = req.body;
 
     const updatedVolunteer = await volunteerModel.shiftCheckIn(volunteerID, scheduleID, shiftDate);
