--- backend/src/models/post_request.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>POST Request</title>
-</head>
-<body>
-    <h1>Send POST Request</h1>
-    <button id="sendRequest">Send Request</button>
-
-    <script>
-        document.getElementById("sendRequest").addEventListener("click", async () => {
-            const url = "http://localhost:3001/volunteer/volunteer/shiftCheckIn";
-            const body = {
-                "volunteerID": "1230545b-0505-4909-826c-59359503dae6",
-                "scheduleID": 100,
-                "shiftDate": "2024-10-17"
-            };
-
-            try {
-                const response = await fetch(url, {
-                    method: "POST",
-                    headers: {
-                        "Content-Type": "application/json"
-                    },
-                    body: JSON.stringify(body)
-                });
-
-                const data = await response.json();
-                console.log(data);
-                alert("Request successful! Check console for response.");
-            } catch (error) {
-                console.error("Error:", error);
-                alert("An error occurred. Check console for details.");
-            }
-        });
-    </script>
-</body>
-</html>
