+++ frontend/src/api/api.js
@@ -5,4 +5,21 @@ const api = axios.create({
     baseURL: backend
 });
 
+// Attach auth token to each req
+api.interceptors.request.use(
+    (config) => {
+      const token = localStorage.getItem('neuronAuthToken');
+
+      console.log(token)
+      if (token) {
+        config.headers.Authorization = `Bearer ${token}`;
+      }
+
+      return config;
+    },
+    (error) => {
+      return Promise.reject(error);
+    }
+);
+
 export default api;
